backend/requirements.txt
Django==5.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
